requirements.txt
Flask==2.0.1
Flask_Mail==0.9.1
Flask_WTF==0.15.1
Flask_Moment==1.0.2
Jinja2==3.0.1
pymongo==3.12.0
mongoengine==0.23.1
openpyxl==3.0.9
Flask_Migrate==3.1.0
requests==2.26.0
urllib3==1.26.7
flask_mongoengine==1.0.0
Flask_Login==0.5.0
WTForms==2.3.3
itsdangerous==2.0.1
python-decouple==3.4
gunicorn==20.1.0
qrcode==7.3.1
PyJWT==2.3.0
PyMySQL==1.0.2
qrcode==7.3.1
Pillow==8.4.0
six==1.16.0
requests==2.26.0
Flask-Dropzone==1.6.0
pytz==2021.3
Babel==2.9.1
Flask-Babel==2.0.0